+++ packages/utils/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.7-alpha.1](https://github.com/selfcommunity/community-js/compare/@selfcommunity/utils@0.2.7-alpha.0...@selfcommunity/utils@0.2.7-alpha.1) (2024-02-02)
+
+
+### Bug Fixes
+
+* categories ordered ([acd4b78](https://github.com/selfcommunity/community-js/commit/acd4b78140d37718f35dc221023d3b59fb0e0acd))
+
+
+
+
+
+## [0.2.7-alpha.0](https://github.com/selfcommunity/community-js/compare/@selfcommunity/utils@0.2.6...@selfcommunity/utils@0.2.7-alpha.0) (2024-02-01)
+
+**Note:** Version bump only for package @selfcommunity/utils
+
+
+
+
+
 ## [0.2.6](https://github.com/selfcommunity/community-js/compare/@selfcommunity/utils@0.2.6-alpha.1...@selfcommunity/utils@0.2.6) (2024-02-01)
 
 **Note:** Version bump only for package @selfcommunity/utils